Fix tests and add docstrings
